apps/hyperliquid-recorder/pyproject.toml
[project]
name = "hyperliquid-recorder"
version = "0.1.0"
description = "Hyperliquid L2 snapshot recorder for ClickHouse"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "clickhouse-connect~=0.8.18",
    "grpcio~=1.73.0",
    "grpcio-tools~=1.73.0",
    "loguru~=0.7.3",
    "prometheus-client~=0.23.1",
    "pydantic~=2.11.9",
]

[build-system]
requires = ["setuptools~=80.9", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "mypy~=1.15.0",
    "pytest~=8.4.2",
    "ruff~=0.9.4",
]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
ignore_missing_imports = true
exclude = ["tests/", "scripts/"]

[[tool.mypy.overrides]]
module = ["hyperliquid_recorder.orderbook_pb2", "hyperliquid_recorder.orderbook_pb2_grpc"]
ignore_errors = true

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests", "scripts"]
exclude = ["src/hyperliquid_recorder/orderbook_pb2.py", "src/hyperliquid_recorder/orderbook_pb2_grpc.py"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "SIM",
    "TCH",
    "RUF",
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["hyperliquid_recorder"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
